+++ workspaces/catalog-cards/src/plugin.ts
@@ -1,7 +1,4 @@
-import {
-  createPlugin,
-  createComponentExtension,
-} from '@backstage/core-plugin-api';
+import { createPlugin } from '@backstage/core-plugin-api';
 
 import { rootRouteRef } from './routes';
 
@@ -12,38 +9,8 @@ export const catalogCardsPlugin = createPlugin({
   },
 });
 
-export const CatalogCardsContent = catalogCardsPlugin.provide(
-  createComponentExtension({
-    name: 'CatalogCardsContent',
-    component: {
-      lazy: () =>
-        import('./components/CatalogCardsContent').then(
-          (m) => m.CatalogCardsContent as any,
-        ),
-    },
-  }),
-);
-
-export const CatalogViewToggle = catalogCardsPlugin.provide(
-  createComponentExtension({
-    name: 'CatalogViewToggle',
-    component: {
-      lazy: () =>
-        import('./components/CatalogViewToggle').then(
-          (m) => m.CatalogViewToggle as any,
-        ),
-    },
-  }),
-);
-
-export const CatalogCardGrid = catalogCardsPlugin.provide(
-  createComponentExtension({
-    name: 'CatalogCardGrid',
-    component: {
-      lazy: () =>
-        import('./components/CatalogCardGrid').then(
-          (m) => m.CatalogCardGrid as any,
-        ),
-    },
-  }),
-);
+// Direct component exports - much simpler integration!
+export { CatalogCardsContent } from './components/CatalogCardsContent';
+export { CatalogViewToggle } from './components/CatalogViewToggle';
+export { CatalogCardGrid } from './components/CatalogCardGrid';
+export { EntityCard } from './components/EntityCard';